# Source/Events/Events.Server/Swagger/v1/Swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Events.Server, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Topics": {
      "get": {
        "tags": [
          "Topics"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Topic"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Topic"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Topic"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Topics"
        ],
        "operationId": "DeleteTopic",
        "parameters": [
          {
            "name": "topicName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Topic"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Topic"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Topic"
                }
              }
            }
          }
        }
      }
    },
    "/api/Topics/{id}": {
      "get": {
        "tags": [
          "Topics"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Topic"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Topic"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Topic"
                }
              }
            }
          }
        }
      }
    },
    "/api/Topics/{topicName}": {
      "post": {
        "tags": [
          "Topics"
        ],
        "operationId": "AddMessage",
        "parameters": [
          {
            "name": "topicId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "message",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "topicName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Topics/message": {
      "delete": {
        "tags": [
          "Topics"
        ],
        "operationId": "DeleteMessage",
        "parameters": [
          {
            "name": "topicId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Topic"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Topic"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Topic"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Entry": {
        "type": "object",
        "properties": {
          "createdOn": {
            "type": "string",
            "format": "date-time"
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "entryId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "Topic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time"
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Entry"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter a valid token",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      },
      "ApiKey": {
        "type": "apiKey",
        "name": "ApiKey",
        "in": "header"
      }
    }
  }
}